+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [20160929]
+    - Fixed crash on print AudioDevice & AudioSession on Python3
+    - Fixed GetAllSessions() reliability, refs #1
+
 ## [20160508]
     - Fixed enum requirement
     - Unit tested examples: